feat: add separated tool services
