+++ src/services/index.js
@@ -3,7 +3,9 @@ import { useLayerToolService } from './layerTool';
 import { useOverlayService } from './overlay';
 import { useLayerQueryService } from './layerQuery';
 import { useNodeQueryService } from './nodeQuery';
-import { useDrawToolService, useEraseToolService, useTopToolService, useGlobalEraseToolService, useCutToolService, useSelectService, useDirectionToolService, usePathToolService } from './tools';
+import { useDrawToolService, useEraseToolService, useTopToolService, useCutToolService } from './singleLayerTools';
+import { useSelectService, useDirectionToolService, useGlobalEraseToolService } from './multiLayerTools';
+import { usePathToolService } from './wandTools';
 import { useToolSelectionService } from './toolSelection';
 import { useViewportService } from './viewport';
 import { useStageResizeService } from './stageResize';
