+++ src/services/wandTools.js
@@ -1,0 +1,61 @@
+import { defineStore } from 'pinia';
+import { watch } from 'vue';
+import { useToolSelectionService } from './toolSelection';
+import { useHamiltonianService } from './hamiltonian';
+import { useLayerQueryService } from './layerQuery';
+import { useStore } from '../stores';
+import { CURSOR_STYLE } from '@/constants';
+
+export const usePathToolService = defineStore('pathToolService', () => {
+    const tool = useToolSelectionService();
+    const hamiltonian = useHamiltonianService();
+    const layerQuery = useLayerQueryService();
+    const { nodeTree, nodes, pixels: pixelStore } = useStore();
+
+    watch(() => tool.prepared, async (p) => {
+        if (p !== 'path' || tool.shape !== 'wand' || nodeTree.selectedLayerCount !== 1) return;
+
+        tool.setCursor({ wand: CURSOR_STYLE.WAIT });
+
+        const layerId = nodeTree.selectedLayerIds[0];
+        const allPixels = pixelStore.get(layerId);
+        const paths = await hamiltonian.traverseFree(allPixels);
+
+        tool.setCursor({ wand: CURSOR_STYLE.PATH });
+
+        if (!paths.length) {
+            tool.setPrepared('done');
+            return;
+        }
+
+        const color = nodes.getProperty(layerId, 'color');
+        const name = nodes.getProperty(layerId, 'name');
+        const groupId = nodes.createGroup({ name: `${name} Paths` });
+
+        nodeTree.insert([groupId], layerQuery.lowermost([layerId]), true);
+
+        nodeTree.remove([layerId]);
+        nodes.remove([layerId]);
+        pixelStore.remove([layerId]);
+
+        paths.forEach((path, idx) => {
+            const subGroupId = nodes.createGroup({ name: `Path ${idx + 1}` });
+            nodeTree.append([subGroupId], groupId, false);
+
+            const ids = [];
+            path.forEach((pixel, j) => {
+                const lid = nodes.createLayer({ name: `Pixel ${j + 1}`, color });
+                pixelStore.addPixels(lid, [pixel]);
+                ids.push(lid);
+            });
+            nodeTree.append(ids, subGroupId, false);
+        });
+
+        nodeTree.replaceSelection([groupId]);
+
+        tool.setPrepared('done');
+    });
+
+    return {};
+});
+
